fix: Move Package & Change Mapping URL
- 패키지 변경
- mapping url 변경
- 카테고리 관리 API는 GroupController로 이동
